Refactor GenerateRuntimeGraph into multiple files
